refactor: existant methods and file names
