+++ prisma/schema.prisma
@@ -27,7 +27,7 @@ model Category {
   father_id  Int?
   created_at DateTime  @default(now())
   updated_at DateTime  @updatedAt
-  Products    Product[]
+  products    Product[]
 
   @@map("categories")
 }
