+++ src/domain/usecases/product.usecase.ts
@@ -1,5 +1,5 @@
-import { DatabaseCategoryRepository } from '@/infrastructure/repositories/catogorie.repositorie';
-import { DatabaseProductRepository } from '@/infrastructure/repositories/product.repositorie';
+import { DatabaseCategoryRepository } from '@/infrastructure/repositories/category.repository';
+import { DatabaseProductRepository } from '@/infrastructure/repositories/product.repository';
 import {
   AddProductDto,
   UpdateProductDto,
@@ -14,17 +14,15 @@ export class ProductUseCase {
     private readonly productRepository: DatabaseProductRepository,
     private readonly categoryRepository: DatabaseCategoryRepository,
   ) {}
+
   async create(product: AddProductDto): Promise<Product> {
-    const category = await this.categoryRepository.findById(
-      product.category_id,
-    );
+    await this.categoryRepository.findById(product.category_id);
     return await this.productRepository.insert(product);
   }
   async update(id: number, product: UpdateProductDto): Promise<void> {
     await this.productRepository.findById(id);
     return await this.productRepository.update(id, product);
   }
-
   async delete(id: number): Promise<void> {
     return await this.productRepository.deleteById(id);
   }
