+++ src/presentation/category/category.module.ts
@@ -3,11 +3,17 @@ import { CategoryController } from './category.controller';
 import { Module } from '@nestjs/common';
 import { RepositoriesModule } from '@/infrastructure/repositories/repositories.module';
 import { CategoryUseCase } from '@/domain/usecases/category.usecase';
-import { DatabaseCategoryRepository } from '@/infrastructure/repositories/catogorie.repositorie';
+import { DatabaseCategoryRepository } from '@/infrastructure/repositories/category.repository';
+import { DatabaseProductRepository } from '@/infrastructure/repositories/product.repository';
 
 @Module({
   imports: [RepositoriesModule],
   controllers: [CategoryController],
-  providers: [CategoryUseCase, DatabaseCategoryRepository, PrismaService],
+  providers: [
+    CategoryUseCase,
+    DatabaseCategoryRepository,
+    DatabaseProductRepository,
+    PrismaService,
+  ],
 })
 export class CategoryModule {}
