+++ src/presentation/product/product.module.ts
@@ -3,8 +3,8 @@ import { ProductController } from './product.controller';
 import { Module } from '@nestjs/common';
 import { RepositoriesModule } from '@/infrastructure/repositories/repositories.module';
 import { ProductUseCase } from '@/domain/usecases/product.usecase';
-import { DatabaseProductRepository } from '@/infrastructure/repositories/product.repositorie';
-import { DatabaseCategoryRepository } from '@/infrastructure/repositories/catogorie.repositorie';
+import { DatabaseProductRepository } from '@/infrastructure/repositories/product.repository';
+import { DatabaseCategoryRepository } from '@/infrastructure/repositories/category.repository';
 
 @Module({
   imports: [RepositoriesModule],
