+++ src/presentation/user/user.controller.ts
@@ -11,6 +11,7 @@ import { ApiExtraModels, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { UserUseCase } from '@/domain/usecases/user.usecase';
 import { AddUserDto, UpdateUserDto } from './user.dto';
 import { UserPresenter } from './user.presenter';
+import { Public } from '@/main/decorators/public.decorator';
 
 @Controller('user')
 @ApiTags('user')
@@ -19,6 +20,7 @@ import { UserPresenter } from './user.presenter';
 export class UserController {
   constructor(private readonly userUseCase: UserUseCase) {}
   @Post()
+  @Public()
   async create(@Body() body: AddUserDto): Promise<UserPresenter> {
     const user = await this.userUseCase.create(body);
     return new UserPresenter(user);
